+++ src/About.js
@@ -1,0 +1,20 @@
+import React from "react";
+const About = () => {
+  return (
+    <div className={"about"}>
+      <div className="col-sm-5 text-center container" id="about-content">
+        <p>
+          This project is a collaborative effort from&nbsp;
+          <a href="https://github.com/bears-team-06">Bears Team 6</a> of Chingu
+          Beta Voyage.
+        </p>
+        <p>
+          Visit <a href="https://chingu.io/">Chingu Cohorts website</a> and join
+          a supportive group of life-long learners.
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default About;
