+++ src/App.js
@@ -1,183 +1,28 @@
-import React, { Component } from "react";
-import ReusableTimeConfigurationComponent from "./components/ReusableTimeConfigurationComponent";
-import ReusableButtonComponent from "./components/ReusableButtonComponent";
-import TimerBox from "./components/TimerBox";
-import TimerState from "./TimerState";
-import DefaultStates from "./DefaultState";
-import PomodoroState from "./PomodoroState";
+import React  from "react";
+import PomodoroClock from "./PomodoroClock";
+import About from "./About";
+import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 
-class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = DefaultStates;
-  }
+const App = () => {
+  return (
+    <Router>
+      <div>
+        <nav>
+          <ul>
+            <li>
+              <Link to="/">Pomodoro Clock</Link>
+            </li>
+            <li>
+              <Link to="/about/">About</Link>
+            </li>
+          </ul>
+        </nav>
 
-  set timerState(state) { // do we even need this property here, it does not do anything other than set/get other property?
-      // this is the place to decide what to do with timers
-      switch (state) {
-          case TimerState.Running:
-              console.log('Start running');
-              break;
-          case TimerState.Stopped:
-               console.log('Stop Running'); // reset session to 1
-              break;
-          case TimerState.Paused:
-              console.log('Pause timer');
-              break;
-          default:
-              throw new Error('Where did you get this state?')
-      }
-      this.setState({
-          'timerState': state
-      });
-  }
-
-  get timerState() {
-    return this.state.timerState;
-  }
-
-  getPomodoroState = sessionNumber => {
-    switch (true) {
-      case sessionNumber % 2 !== 0:
-        return PomodoroState.Focus;
-      case sessionNumber % 6 === 0:
-        return PomodoroState.LongBreak;
-      case sessionNumber % 2 === 0:
-        return PomodoroState.ShortBreak;
-      default:
-        throw new Error("Unknown Pomodoro State");
-    }
-  };
-
-  onTimerComplete = () => {
-    this.timerState = TimerState.Paused;
-    const incrementedState = this.state.sessionNumber + 1;
-    const nextPomodoroState = this.getPomodoroState(incrementedState);
-    this.setState({
-      sessionNumber: incrementedState,
-      pomodoroState: nextPomodoroState
-    });
-  };
-
-  get sessionName() {
-    const currentPomodoro = this.state.pomodoroState;
-    switch (currentPomodoro) {
-      case PomodoroState.Focus:
-        return "Focus";
-      case PomodoroState.ShortBreak:
-        return "Short Break";
-      case PomodoroState.LongBreak:
-        return "Long Break";
-      default:
-        return "";
-    }
-  }
-
-  get sessionTime() {
-    const currentPomodoro = this.state.pomodoroState;
-    switch (currentPomodoro) {
-      case PomodoroState.Focus:
-        return this.state.focusTime;
-      case PomodoroState.ShortBreak:
-        return this.state.shortBreakTime;
-      case PomodoroState.LongBreak:
-        return this.state.longBreakTime;
-      default:
-        return 0;
-    }
-  }
-
-  get isStartButtonDisabled() {
-    return this.state.timerState === TimerState.Running
-  }
-
-  startButtonClickHandler = () => {
-    this.timerState = TimerState.Running;
-  };
-
-  get isPauseButtonDisabled() {
-    return this.state.timerState !== TimerState.Running || this.state.timerState === TimerState.Stopped;
-  }
-
-  pauseButtonClickHandler = () => {
-    this.timerState = TimerState.Paused;
-  };
-
-  get isResetButtonDisabled() {
-      return false
-  }
-
-  resetButtonClickHandler = () => {
-      this.setState(DefaultStates)
-  }
-
-  onShortBreakTimeChange = (newValue) => {
-      this.setState({
-          shortBreakTime: newValue
-      })
-  }
-
-  onSessionTimeChange = (newValue) => {
-    this.setState({
-      focusTime: newValue
-    })
-  }
-
-  render() {
-    return (
-
-      <div className={"main-body"}>
-        <div className={"row time-configuration"}>
-          <ReusableTimeConfigurationComponent
-            className = "Break-Length"
-            labelName="Break Length"
-            timeLength={this.state.shortBreakTime} 
-            onChange={this.onShortBreakTimeChange}
-            minimumChange={60}
-            maximumLength={3600}
-            minimumLength={60}
-          />
-          <ReusableTimeConfigurationComponent 
-            labelName='Session Length'
-            className = "Session-Length"
-            timeLength={this.state.focusTime} 
-            onChange={this.onSessionTimeChange} 
-            minimumChange={300} 
-            maximumLength={7200} 
-            minimumLength={600} 
-          />
-        </div>
-        <div className={"row timer-box"}>
-            <TimerBox
-              sessionName={this.sessionName}
-              sessionTime={this.sessionTime}
-              timerState={this.state.timerState}
-              onTimerComplete={this.onTimerComplete}
-            />
-        </div>
-        <div className={"row button-container"}>
-          <ReusableButtonComponent
-            className={"Start"}
-            label={"Start"}
-            isDisabled={this.isStartButtonDisabled}
-            clickHandler={this.startButtonClickHandler}
-          />
-          <ReusableButtonComponent
-            className={"Pause"}
-            label={"Pause"}
-            isDisabled={this.isPauseButtonDisabled}
-            clickHandler={this.pauseButtonClickHandler}
-          />
-          <ReusableButtonComponent
-            className={"Stop"}
-            label={"Stop"}
-            isDisabled={this.isResetButtonDisabled}
-            clickHandler={this.resetButtonClickHandler}
-          />
-        </div>
+        <Route path="/" exact component={PomodoroClock} />
+        <Route path="/about/" component={About} />
       </div>
-    );
-  }
-}
+    </Router>
+  );
+};
 
 export default App;
